backend/requirements.txt
fastapi
uvicorn[standard]
sse-starlette
openai
pydantic
python-dotenv
httpx